pyproject.toml
[project]
name = "OpenAgent"
version = "0.1.0"
description = "Agents & MCP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "langfuse>=4.3.1",
    "mcp[cli]>=1.6.0",
    "openai>=1.73.0",
    "openai-agents==0.12.5",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.0",
]


[tool.ruff]
# Python version (affects how Ruff parses and understands syntax)
target-version = "py313"
# Set max line length — 88 is safe, Black-compatible
line-length = 88

[tool.ruff.lint]
# Enable important rule categories
select = [
"E",    # pycodestyle (style issues)
"F",    # pyflakes (logical issues)
"I",    # isort (import sorting)              Keeps import sections clean and sorted
"C90",  # mccabe (complexity checks)          Warns if an endpoint function is too complex
"B",    # bugbear (common mistakes)           Flags common Python bugs (e.g., using mutable default args)
"ASYNC",# asyncio checks (for FastAPI)        Ensures you're using await correctly in endpoints
"A",    # flake8-builtins (naming conflicts)  Catches accidental use of names like list, dict, etc
"Q",    # flake8-quotes (quote consistency)   Ensures consistent use of ' vs " in strings
]

# Ignore specific rules that conflict with Black or are too noisy
ignore = [
"E203",  # Whitespace before ':' (Black handles this)
"E501",  # Line too long (Black handles wrapping)
"F401",  # Unused imports — disable this if using __init__.py with imports
]

# Enable import sorting with Ruff (replaces isort)
[tool.ruff.lint.isort]
known-first-party = ["app"]  # replace "app" with your local FastAPI package name
combine-as-imports = true
force-single-line = false
split-on-trailing-comma = false


# Formatting options (optional)
[tool.ruff.format]
quote-style = "double"        # consistent double quotes
indent-style = "space"
skip-magic-trailing-comma = true
